provide a `hash` method for `QabElem`
- moved `minimize` and `conductor` from `experimental/GModule/GModule.jl` to `src/Rings/AbelianClosure.jl`
- added `hash(a::QabElem, h::UInt)`

resolves #1287
